refactor(faction-detail): dispatch the page body, not just the hero
The hero already dispatched via FACTION_HEROES; the members/tasks/recent-praxis
sections were uniform placeholder. Splits the fetch into useFactionDetail (the
stable FactionDetailState contract + the per-faction backdrop side-effect) and
adds a second FACTION_BODY dispatcher (via pickVariant) with DefaultFactionBody
holding the original sections verbatim, so any faction can restyle its body by
registering one map entry. No faction registers a bespoke body yet, so every
faction renders the Default path unchanged.
